+++ renderer/src/main.ts
@@ -4,12 +4,14 @@ import * as I18n from './web/i18n'
 import * as Data from './assets/data'
 import { initConfig, AppConfig } from './web/Config'
 import { Host } from './web/background/IPC'
+import { installNativeNumberInputCapture } from './web/native-number-input'
 
 ;(async function () {
   await initConfig()
   const i18nPlugin = await I18n.init(AppConfig().language)
   await Data.init(AppConfig().language)
   await Host.init()
+  installNativeNumberInputCapture()
 
   watch(() => AppConfig().language, async () => {
     await Data.loadForLang(AppConfig().language)