+++ renderer/src/web/price-check/PriceCheckWindow.vue
@@ -8,7 +8,9 @@
     <div v-if="!isBrowserShown" class="layout-column shrink-0"
       style="width: var(--game-panel);">
     </div>
-    <div id="price-window" class="layout-column shrink-0 text-gray-200 pointer-events-auto" style="width: 28.75rem;">
+    <div v-if="!isBrowserShown" id="price-window"
+      class="layout-column shrink-0 text-gray-200 pointer-events-auto"
+      style="width: 28.75rem;">
       <AppTitleBar @close="closePriceCheck" @click="openLeagueSelection" :title="title">
         <ui-popover v-if="stableOrbCost" trigger="click" boundary="#price-window">
           <template #target>